backend/requirements.txt
Django==3.0
django-cors-headers==3.2.0
django-extensions==2.2.5
djangorestframework==3.10.3
psycopg2==2.8.4

